support git folder to be inside documentation folder
